requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.10.0
WTForms==3.1.1
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4
python-dateutil==2.8.2
gunicorn==21.2.0
Werkzeug==3.0.1
email-validator==2.1.0
pdfplumber==0.11.9
Pillow==12.1.1
